Cast statvfs fields to u64 when calculating free space
Fixes a compile error on MacOS where `statvfs.f_bavail` and `statvfs.f_frsize`
are `u32` (or simply smaller integers than `u64`), causing a mismatched types
error when multiplying and returning as `u64`. Casting both operands to `u64`
before multiplication fixes the issue.

+++ src/lib/builder.rs
@@ -425,7 +425,7 @@ impl Builder {
         if ret != 0 {
             return Err(std::io::Error::last_os_error());
         }
-        Ok(stat.f_bavail * stat.f_frsize)
+        Ok((stat.f_bavail as u64) * (stat.f_frsize as u64))
     }
 
     fn process_file(&mut self, path: PathBuf) -> Result<bool> {
